refactor: Extract diagnostic as crate
