+++ crates/diagnostic/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "diagnostic"
+version = { workspace = true }
+edition = { workspace = true }
+license = { workspace = true }
+
+[lints]
+workspace = true
+
+[dependencies]
+ast = { workspace = true }
+base_db = { workspace = true }
+lsp = { workspace = true }
+hir = { workspace = true }
+hir_ty = { workspace = true }
