+++ crates/diagnostic/src/lib.rs
@@ -1,0 +1,480 @@
+//! Define the content of error output for users.
+//! For example, used in LSP and Dock.
+
+use ast::AstNode;
+
+/// 診断情報
+/// 基本的には、1つのエラーから1つの診断情報が作成されます。そのため、同じファイル
+/// 場合によっては、1つのエラーから1つの診断情報の中に複数のメッセージが含まれる場合があります。
+#[derive(Debug)]
+pub struct Diagnostic {
+    /// 診断情報対象のファイル
+    pub file: hir::NailFile,
+    /// 診断情報のタイトル(サマリ)
+    pub title: String,
+    /// 診断情報の開始位置
+    /// メッセージ一覧内の中で一番小さい開始位置です。
+    pub head_offset: usize,
+    /// 診断情報のメッセージ一覧
+    pub messages: Vec<Message>,
+}
+impl Diagnostic {
+    /// hir_tyの型推論エラー情報を元に診断情報を作成します。
+    pub fn from_hir_ty_inference_error(
+        db: &base_db::SalsaDatabase,
+        file_db: &hir::HirFileDatabase,
+        source_map: &hir::HirFileSourceMap,
+        error: &hir_ty::InferenceError,
+    ) -> Self {
+        use hir_ty::InferenceError;
+        let file = source_map.file(db);
+
+        match error {
+            InferenceError::MismatchedTypeIfCondition {
+                expected_condition_bool_ty,
+                found_condition_ty,
+                found_condition_expr,
+            } => {
+                let text_range = found_condition_expr.text_range(db, source_map);
+                let cond_type = type_to_string(db, expected_condition_bool_ty);
+                let found_cond_ty = type_to_string(db, found_condition_ty);
+
+                Diagnostic {
+                    file,
+                    title: "Mismatched type in if condition".to_string(),
+                    head_offset: text_range.start().into(),
+                    messages: vec![Message {
+                        message: format!("expected {cond_type}, actual: {found_cond_ty}"),
+                        range: (text_range.start().into())..(text_range.end().into()),
+                    }],
+                }
+            }
+            InferenceError::MismatchedTypeElseBranch {
+                then_branch_ty,
+                then_branch,
+                else_branch_ty,
+                else_branch,
+            } => {
+                let then_branch_range = then_branch.text_range(db, source_map);
+                let else_branch_range = else_branch.text_range(db, source_map);
+                let hir::Expr::Block(then_block) = then_branch.lookup(file_db) else {
+                    unreachable!()
+                };
+                let then_tail_range: std::ops::Range<usize> = if let Some(tail) = then_block.tail {
+                    tail.text_range(db, source_map).into()
+                } else {
+                    then_branch_range.end().into()..then_branch_range.end().into()
+                };
+                let hir::Expr::Block(else_block) = else_branch.lookup(file_db) else {
+                    unreachable!()
+                };
+                let else_tail_range: std::ops::Range<usize> = if let Some(tail) = else_block.tail {
+                    tail.text_range(db, source_map).into()
+                } else {
+                    else_branch_range.end().into()..else_branch_range.end().into()
+                };
+
+                let then_branch_ty = type_to_string(db, then_branch_ty);
+                let else_branch_ty = type_to_string(db, else_branch_ty);
+
+                Diagnostic {
+                    file,
+                    title: "Mismatched type if branch and else branch".to_string(),
+                    head_offset: then_branch_range.start().into(),
+                    messages: vec![
+                        Message {
+                            message: format!("expected {then_branch_ty}, actual: {else_branch_ty}"),
+                            range: (then_branch_range.start().into())
+                                ..(else_branch_range.start().into()),
+                        },
+                        Message {
+                            range: then_tail_range,
+                            message: format!("Type is {then_branch_ty}"),
+                        },
+                        Message {
+                            range: else_tail_range,
+                            message: format!("Type is {else_branch_ty}"),
+                        },
+                    ],
+                }
+            }
+            InferenceError::MismatchedTypeOnlyIfBranch {
+                then_branch_ty,
+                then_branch,
+                else_branch_unit_ty,
+            } => {
+                let then_branch_range = then_branch.text_range(db, source_map);
+                let hir::Expr::Block(then_block) = then_branch.lookup(file_db) else {
+                    unreachable!()
+                };
+                let then_tail_range: std::ops::Range<usize> = if let Some(tail) = then_block.tail {
+                    tail.text_range(db, source_map).into()
+                } else {
+                    unreachable!("末尾の式がない場合は必ずUnitなのでエラーとならないはずです。");
+                };
+
+                let then_branch_ty = type_to_string(db, then_branch_ty);
+                let else_branch_ty = type_to_string(db, else_branch_unit_ty);
+
+                Diagnostic {
+                    file,
+                    title: "Mismatched type if branch and else branch".to_string(),
+                    head_offset: then_branch_range.start().into(),
+                    messages: vec![
+                        Message {
+                            message: format!("expected {else_branch_ty}, actual {then_branch_ty}"),
+                            range: (then_branch_range.start().into())
+                                ..(then_branch_range.end().into()),
+                        },
+                        Message {
+                            range: then_tail_range,
+                            message: format!("Type is {then_branch_ty}"),
+                        },
+                    ],
+                }
+            }
+            InferenceError::MismaatchedTypeCallArg {
+                expected_ty,
+                found_ty,
+                expected_signature: _,
+                found_expr,
+                arg_pos: _,
+            } => {
+                let text_range = found_expr.text_range(db, source_map);
+                let expected_ty = type_to_string(db, expected_ty);
+                let found_ty = type_to_string(db, found_ty);
+
+                Diagnostic {
+                    file,
+                    title: "Mismatched type in call argument".to_string(),
+                    head_offset: text_range.start().into(),
+                    messages: vec![Message {
+                        message: format!("expected {expected_ty}, actual: {found_ty}"),
+                        range: (text_range.start().into())..(text_range.end().into()),
+                    }],
+                }
+            }
+            InferenceError::MismatchedBinaryInteger {
+                expected_int_ty,
+                found_expr,
+                found_ty,
+                op,
+            } => {
+                let text_range = found_expr.text_range(db, source_map);
+                let expected_ty = type_to_string(db, expected_int_ty);
+                let found_ty = type_to_string(db, found_ty);
+
+                let op = match op {
+                    ast::BinaryOp::Add(_) => "+",
+                    ast::BinaryOp::Sub(_) => "-",
+                    ast::BinaryOp::Mul(_) => "*",
+                    ast::BinaryOp::Div(_) => "/",
+                    ast::BinaryOp::Equal(_)
+                    | ast::BinaryOp::GreaterThan(_)
+                    | ast::BinaryOp::LessThan(_)
+                    | ast::BinaryOp::Assign(_) => unreachable!(),
+                };
+
+                Diagnostic {
+                    file,
+                    title: format!("Mismatched type in binary integer expression `{op}`"),
+                    head_offset: text_range.start().into(),
+                    messages: vec![Message {
+                        message: format!("expected {expected_ty}, actual: {found_ty}"),
+                        range: (text_range.start().into())..(text_range.end().into()),
+                    }],
+                }
+            }
+            InferenceError::MismatchedBinaryCompare {
+                compare_from_ty,
+                compare_from_expr,
+                compare_to_ty,
+                compare_to_expr,
+                op,
+            } => {
+                let compare_from_range = compare_from_expr.text_range(db, source_map);
+                let compare_to_range = compare_to_expr.text_range(db, source_map);
+                let compare_from_ty = type_to_string(db, compare_from_ty);
+                let compare_to_ty = type_to_string(db, compare_to_ty);
+
+                let op = match op {
+                    ast::BinaryOp::Equal(_) => "==",
+                    ast::BinaryOp::GreaterThan(_) => ">",
+                    ast::BinaryOp::LessThan(_) => "<",
+                    ast::BinaryOp::Add(_)
+                    | ast::BinaryOp::Sub(_)
+                    | ast::BinaryOp::Mul(_)
+                    | ast::BinaryOp::Div(_)
+                    | ast::BinaryOp::Assign(_) => unreachable!(),
+                };
+
+                Diagnostic {
+                    file,
+                    title: format!("Mismatched type in binary compare expression `{op}`"),
+                    head_offset: compare_from_range.start().into(),
+                    messages: vec![
+                        Message {
+                            range: compare_from_range.into(),
+                            message: format!("Type is {compare_from_ty}"),
+                        },
+                        Message {
+                            range: compare_to_range.into(),
+                            message: format!("Type is {compare_to_ty}"),
+                        },
+                    ],
+                }
+            }
+            InferenceError::MismatchedUnary {
+                expected_ty,
+                found_ty,
+                found_expr,
+                op,
+            } => {
+                let text_range = found_expr.text_range(db, source_map);
+                let expected_ty = type_to_string(db, expected_ty);
+                let found_ty = type_to_string(db, found_ty);
+
+                let op = match op {
+                    ast::UnaryOp::Not(_) => "!",
+                    ast::UnaryOp::Neg(_) => "-",
+                };
+
+                Diagnostic {
+                    file,
+                    title: format!("Mismatched type in unary expression `{op}`"),
+                    head_offset: text_range.start().into(),
+                    messages: vec![Message {
+                        message: format!("expected {expected_ty}, actual: {found_ty}"),
+                        range: (text_range.start().into())..(text_range.end().into()),
+                    }],
+                }
+            }
+            InferenceError::MismatchedTypeReturnExpr {
+                expected_signature,
+                found_ty,
+                found_return_expr: _,
+                found_return,
+            } => {
+                let text_range = found_return.text_range(db, source_map);
+                let return_type = expected_signature.return_type(db);
+                let return_type = type_to_string(db, &return_type);
+                let found_ty = type_to_string(db, found_ty);
+
+                Diagnostic {
+                    file,
+                    title: "Mismatched type in return expression".to_string(),
+                    head_offset: text_range.start().into(),
+                    messages: vec![Message {
+                        message: format!("expected {return_type}, actual: {found_ty}"),
+                        range: (text_range.start().into())..(text_range.end().into()),
+                    }],
+                }
+            }
+            InferenceError::MismatchedTypeReturnValue {
+                expected_signature,
+                expected_function,
+                found_ty,
+                found_last_expr,
+            } => {
+                if let Some(found_last_expr) = found_last_expr {
+                    let text_range = found_last_expr.text_range(db, source_map);
+                    let return_type = expected_signature.return_type(db);
+                    let return_type = type_to_string(db, &return_type);
+                    let found_ty = type_to_string(db, found_ty);
+
+                    Diagnostic {
+                        file,
+                        title: "Mismatched type in return value".to_string(),
+                        head_offset: text_range.start().into(),
+                        messages: vec![Message {
+                            message: format!("expected {return_type}, actual: {found_ty}"),
+                            range: (text_range.start().into())..(text_range.end().into()),
+                        }],
+                    }
+                } else {
+                    let text_range = {
+                        source_map
+                            .source_by_function(db)
+                            .get(expected_function)
+                            .unwrap()
+                            .value
+                            .return_type()
+                            .unwrap()
+                            .syntax()
+                            .text_range()
+                    };
+
+                    let return_type = expected_signature.return_type(db);
+                    let return_type = type_to_string(db, &return_type);
+                    let found_ty = type_to_string(db, found_ty);
+
+                    Diagnostic {
+                        file,
+                        title: "Mismatched type in return value".to_string(),
+                        head_offset: text_range.start().into(),
+                        messages: vec![Message {
+                            message: format!(
+                                "expected {return_type}, actual: {found_ty} as its body has tail"
+                            ),
+                            range: (text_range.start().into())..(text_range.end().into()),
+                        }],
+                    }
+                }
+            }
+            InferenceError::MismatchedCallArgCount {
+                expected_callee_arg_count,
+                found_arg_count,
+                found_expr,
+            } => {
+                let text_range = source_map
+                    .source_by_expr(db)
+                    .get(found_expr)
+                    .unwrap()
+                    .value
+                    .node
+                    .text_range();
+                let expected_callee_arg_count = *expected_callee_arg_count;
+                let found_arg_count = *found_arg_count;
+
+                Diagnostic {
+                    file,
+                    title: "Mismatched argument count in call".to_string(),
+                    head_offset: text_range.start().into(),
+                    messages: vec![Message {
+                        message: format!(
+                            "expected {expected_callee_arg_count} argument, found: {found_arg_count}"
+                        ),
+                        range: (text_range.start().into())..(text_range.end().into()),
+                    }],
+                }
+            }
+            InferenceError::NotCallable {
+                found_callee_ty,
+                found_callee_symbol: _,
+                found_callee_expr,
+            } => {
+                let text_range = found_callee_expr.text_range(db, source_map);
+                let found_callee_ty = type_to_string(db, found_callee_ty);
+
+                Diagnostic {
+                    file,
+                    title: "Not callable".to_string(),
+                    head_offset: text_range.start().into(),
+                    messages: vec![Message {
+                        message: format!("expected <function>, found: {found_callee_ty}"),
+                        range: (text_range.start().into())..(text_range.end().into()),
+                    }],
+                }
+            }
+            InferenceError::ModuleAsExpr {
+                found_module: _,
+                found_expr,
+            } => {
+                let text_range = found_expr.text_range(db, source_map);
+
+                Diagnostic {
+                    file,
+                    title: "Module as expression".to_string(),
+                    head_offset: text_range.start().into(),
+                    messages: vec![Message {
+                        message: "expected <expression>".to_string(),
+                        range: (text_range.start().into())..(text_range.end().into()),
+                    }],
+                }
+            }
+            InferenceError::MismatchedType {
+                expected_ty,
+                expected_expr,
+                found_ty,
+                found_expr,
+            } => {
+                let found_expr_text_range = found_expr.text_range(db, source_map);
+                let expected_ty = type_to_string(db, expected_ty);
+                let found_ty = type_to_string(db, found_ty);
+
+                if let Some(expected_expr) = expected_expr {
+                    let expected_expr_text_range = expected_expr.text_range(db, source_map);
+                    Diagnostic {
+                        file,
+                        title: "Mismatched type".to_string(),
+                        head_offset: expected_expr_text_range.start().into(),
+                        messages: vec![
+                            Message {
+                                message: format!("Type is {expected_ty}"),
+                                range: (expected_expr_text_range.start().into())
+                                    ..(expected_expr_text_range.end().into()),
+                            },
+                            Message {
+                                message: format!("Type is {found_ty}"),
+                                range: (found_expr_text_range.start().into())
+                                    ..(found_expr_text_range.end().into()),
+                            },
+                        ],
+                    }
+                } else {
+                    Diagnostic {
+                        file,
+                        title: "Mismatched type".to_string(),
+                        head_offset: found_expr_text_range.start().into(),
+                        messages: vec![Message {
+                            message: format!("expected {expected_ty}, actual: {found_ty}"),
+                            range: (found_expr_text_range.start().into())
+                                ..(found_expr_text_range.end().into()),
+                        }],
+                    }
+                }
+            }
+            InferenceError::BreakOutsideOfLoop { kind, found_expr } => {
+                let text_range = found_expr.text_range(db, source_map);
+                let kind_text = match kind {
+                    hir_ty::BreakKind::Break => "Break",
+                    hir_ty::BreakKind::Continue => "Continue",
+                };
+
+                Diagnostic {
+                    file,
+                    title: format!("{kind_text} outside of loop"),
+                    head_offset: text_range.start().into(),
+                    messages: vec![Message {
+                        message: "expected in <loop>".to_string(),
+                        range: (text_range.start().into())..(text_range.end().into()),
+                    }],
+                }
+            }
+        }
+    }
+}
+
+/// 診断情報のメッセージ
+#[derive(Debug)]
+pub struct Message {
+    /// メッセージの対象範囲
+    pub range: std::ops::Range<usize>,
+    /// メッセージの内容
+    pub message: String,
+}
+
+fn type_to_string(db: &base_db::SalsaDatabase, ty: &hir_ty::Monotype) -> String {
+    match ty {
+        hir_ty::Monotype::Integer => "int".to_string(),
+        hir_ty::Monotype::Bool => "bool".to_string(),
+        hir_ty::Monotype::Unit => "()".to_string(),
+        hir_ty::Monotype::Function(signature) => {
+            format!(
+                "{} -> {}",
+                signature
+                    .params(db)
+                    .iter()
+                    .map(|param| type_to_string(db, param))
+                    .collect::<Vec<String>>()
+                    .join(", "),
+                type_to_string(db, &signature.return_type(db))
+            )
+        }
+        hir_ty::Monotype::Char => "char".to_string(),
+        hir_ty::Monotype::String => "string".to_string(),
+        hir_ty::Monotype::Variable(_) => unreachable!(),
+        hir_ty::Monotype::Never => "!".to_string(),
+        hir_ty::Monotype::Unknown => "<unknown>".to_string(),
+    }
+}
